fix test server
