+++ src/index.ts
@@ -1,34 +1,8 @@
 import { Response, NextFunction, Request } from 'express'
-import cookieSession from 'cookie-session'
 import { compose } from 'compose-middleware'
 import Logger from 'blgr'
 import { node, invoice, parseEnv, paywall, validateLsat } from './routes'
-import { getEnvVars } from './helpers'
 import { BoltwallConfig, LoggerInterface } from './typings'
-
-const { SESSION_SECRET } = getEnvVars()
-
-const keys = SESSION_SECRET ? [SESSION_SECRET] : []
-
-// a session cookie to store request macaroons in
-const rootMacaroon = cookieSession({
-  name: 'macaroon',
-  maxAge: 86400000,
-  secret: SESSION_SECRET,
-  keys,
-  overwrite: true,
-  signed: true,
-})
-
-// separate cookie for the discharge macaroon
-const dischargeMacaroon = cookieSession({
-  name: 'dischargeMacaroon',
-  maxAge: 86400000,
-  secret: SESSION_SECRET,
-  keys,
-  overwrite: true,
-  signed: true,
-})
 
 function errorHandler(
   err: any, // eslint-disable-line @typescript-eslint/no-explicit-any
@@ -66,8 +40,6 @@ export function boltwall(
     req.boltwallConfig = config
     return compose([
       parseEnv,
-      rootMacaroon,
-      dischargeMacaroon,
       node,
       invoice,
       validateLsat,
